+++ pkgs/api_summary/lib/src/extensions.dart
@@ -23,10 +23,6 @@ extension FormalParameterElementExtension on FormalParameterElement {
       metadata.hasDeprecated;
 }
 
-extension ElementAnnotationListExtension on List<ElementAnnotation> {
-  bool get hasVisibleForTesting => any((a) => a.isVisibleForTesting);
-}
-
 extension IterableIterableExtension on Iterable<Iterable<Object?>> {
   /// Forms a list containing [prefix], followed by the elements of `this`
   /// (separated by [separator]), followed by [suffix].