+++ pkgs/coverage/test/filter_ignored_test.dart
@@ -22,9 +22,7 @@ void main() {
         {1: 'abc', 2: 'def'},
         {1: 1, 2: 2, 3: 3},
       ),
-      fileUri('another_nonexistent_file.dart'): HitMap(
-        {1: 1, 2: 2, 3: 3},
-      ),
+      fileUri('another_nonexistent_file.dart'): HitMap({1: 1, 2: 2, 3: 3}),
       fileUri('test/test_files/test_app.dart'): HitMap(
         {1: 1, 2: 2, 3: 3},
         {1: 1, 2: 2, 3: 3},
@@ -41,14 +39,12 @@ void main() {
 
     // Lines ignored in test/test_files/test_app_isolate.dart.
     const ignores = [
-      52,
       54,
-      55,
       56,
       57,
       58,
-      63,
-      64,
+      59,
+      60,
       65,
       66,
       67,
@@ -58,6 +54,8 @@ void main() {
       71,
       72,
       73,
+      74,
+      75,
     ];
 
     final expected = {
@@ -67,30 +65,30 @@ void main() {
         {1: 'abc', 2: 'def'},
         {1: 1, 2: 2, 3: 3},
       ),
-      fileUri('another_nonexistent_file.dart'): HitMap(
-        {1: 1, 2: 2, 3: 3},
-      ),
+      fileUri('another_nonexistent_file.dart'): HitMap({1: 1, 2: 2, 3: 3}),
       fileUri('test/test_files/test_app_isolate.dart'): HitMap(
         {
           for (var i = 50; i < 100; ++i)
-            if (!ignores.contains(i)) i: i
+            if (!ignores.contains(i)) i: i,
         },
         {
           for (var i = 50; i < 100; ++i)
-            if (!ignores.contains(i)) i: i
+            if (!ignores.contains(i)) i: i,
         },
         {for (var i = 50; i < 100; ++i) i: '$i'},
         {
           for (var i = 50; i < 100; ++i)
-            if (!ignores.contains(i)) i: i
+            if (!ignores.contains(i)) i: i,
         },
       ),
     };
 
     final resolver = await Resolver.create(packagePath: '.');
 
-    final actual =
-        hitmaps.filterIgnored(ignoredLinesInFilesCache: {}, resolver: resolver);
+    final actual = hitmaps.filterIgnored(
+      ignoredLinesInFilesCache: {},
+      resolver: resolver,
+    );
 
     expect(actual.keys.toList(), expected.keys.toList());
     for (final source in expected.keys) {