+++ pkgs/package_config/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 3.0.1-wip
+
 ## 3.0.0
 
 - Adds discovery API to find both a configuration and its location: